Add a `Reserved Regions` entry in the ini
Every Profile can have this field after the `Palette Table Pointers
Address` in order to specify which areas of the ROM must not be used
from OWM.

These ranges will be a list of {start addr}-{end addr}. Under the hood
when checking for free space OWM will be also checking if the address is
not in these regions.

To efficiently make this check there will be one more array, with size
equal to the original ROM, that will contain which bytes should not be
used from these regions.